% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{csvread}
\alias{csvread}
\alias{csvread-package}
\title{Utilities for fast load of CSV files and timing.}
\description{
Various R utilities, some having to do with loading CSV files, Netezza access, timing.
\tabular{ll}{
Package: \tab csvread\cr
Type: \tab Package\cr
Version: \tab 3.5\cr
Date: \tab 2014-05-31\cr
License: \tab Apache License, Version 2.0\cr
Author: \tab Sergei Izrailev, Jeremy Stanley\cr
Maintainer: \tab Sergei Izrailev <sizrailev@collective.com>\cr
LazyLoad: \tab yes\cr
}
}
\details{
This package contains functions for fast loading CSV files,
and convenient timing functions.

\code{\link{csvread}} is a CSV file loader that is much faster than the built-in read.csv,
but needs column type specifications.

See also \code{\link{int64}} for information about dealing with 64-bit integers when loading data from
CSV files.

\code{\link{cm.tic}} provides the timing functions \code{tic} and \code{toc} that can be nested.
One can record all timings while a complex script is running, and examine the values later.

A stack implemented as a vector (\code{\link{cm.stack}}) and as a list (\code{\link{cm.list}})
with push, pop, first, last and clear operations are implemented.

There are also a few other utilities listed under \code{link{cm.utils}}.
}
\author{
Sergei Izrailev, Jeremy Stanley
}
\keyword{64-bit}
\keyword{bigint}
\keyword{csv}
\keyword{profiling}
\keyword{timing}

