% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Ops.int64}
\alias{+}
\alias{+.int64}
\alias{-}
\alias{-.int64}
\alias{<}
\alias{Ops.int64}
\title{Operators for the \code{int64} class.}
\usage{
e1 + e2
e1 - e2

\method{+}{int64}(e1, e2)

\method{-}{int64}(e1, e2)
}
\arguments{
\item{e1}{int64 object, character vector or numeric vector
(character and numeric values are converted by \code{as.int64}).}

\item{e2}{int64 object, character vector or numeric vector
(character and numeric values are converted by \code{as.int64}).}
}
\description{
Operators for the \code{int64} class: one of
\code{+}, \code{-}, \code{==}, \code{!=}, \code{<}, \code{<=}, \code{>} or \code{>=}.
}
\seealso{
int64
}

